refactor(sql): move transaction refresh helpers
