+++ tests/test_api_institutions.py
@@ -49,12 +49,12 @@ sys.modules["app.utils.finance_utils"] = finance_utils_stub
 
 # SQL stub
 sql_pkg = types.ModuleType("app.sql")
-account_logic_stub = types.ModuleType("app.sql.account_logic")
-account_logic_stub.refresh_data_for_plaid_account = lambda *a, **k: True
-account_logic_stub.refresh_data_for_teller_account = lambda *a, **k: True
+transactions_logic_stub = types.ModuleType("app.sql.transactions_logic")
+transactions_logic_stub.refresh_data_for_plaid_account = lambda *a, **k: True
+transactions_logic_stub.refresh_data_for_teller_account = lambda *a, **k: True
 sys.modules["app.sql"] = sql_pkg
-sys.modules["app.sql.account_logic"] = account_logic_stub
-sql_pkg.account_logic = account_logic_stub
+sys.modules["app.sql.transactions_logic"] = transactions_logic_stub
+sql_pkg.transactions_logic = transactions_logic_stub
 
 # Models stub
 models_stub = types.ModuleType("app.models")
